New Changes for Control Algorithm
I changed the pipeline for how we deal with graph generation, and the control algorithm needed a change to match this framework.
Now, all data about initial conditions (including the parameter structure) are saved within the file holding each graph. In addition, we also hold the indices of valid targets and the stable state reached when a given target is made extinct.