Changes to Model Stat Tests
I just added breakdown of the multiple extinction statistic. Namely, whether the multiple extinction includes the target species or not. If there are multiple extinctions that include the target, it's likely that we have too greedy an algorithm and are enforcing a scorched earth policy where we don't care who dies as long as the one we want to die dies too.