Generate Different Values of h and B0
This script will generate graphs with different h and B0 values, based on a given range.
To avoid long computation times, avoid being too resolute. Smaller values of h and B0 tend to have less viable graph topologies, so they take longer to make. A good range for b0 is 0.1 and above, and for h stay within 1.0 and 2.0 to remain valid in an ecological sense.